Restructure web module to include JS code for multithreaded and non-multithreaded
