+++ web/lib/src/sync/multithreaded-pool.worker.ts
@@ -1,7 +1,7 @@
-import init, {sync_worker_entry} from "../wasm/maplibre"
+import * as maplibre from "../wasm/maplibre"
 
 onmessage = async message => {
-    const initialised = init(message.data[0], message.data[1]).catch(err => {
+    const initialised = maplibre.default(message.data[0], message.data[1]).catch(err => {
         // Propagate to main `onerror`:
         setTimeout(() => {
             throw err;
@@ -13,6 +13,7 @@ onmessage = async message => {
     self.onmessage = async message => {
         // This will queue further commands up until the module is fully initialised:
         await initialised;
-        await sync_worker_entry(message.data);
+        // @ts-ignore TODO may not exist
+        await maplibre.sync_worker_entry(message.data);
     };
 }
